requirements.txt
asgiref==3.6.0
Django==4.1.6
sqlparse==0.4.3
psycopg2-binary==2.9.5
djangorestframework==3.14.0
django-filter==2.4.0
python-dateutil==2.8.2
